# swagger/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "version" : "v1",
    "title" : "API"
  },
  "host" : "www.mannetroll.com",
  "basePath" : "/eswebbapp",
...
  "schemes" : [ "https" ],
  "paths" : {
...
    "/count/{from}/{to}" : {
      "get" : {
        "tags" : [ "Volume" ],
        "summary" : "count",
        "description" : "count",
        "operationId" : "count",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "from",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "to",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
...
    "/populartimes/spid/{spid}" : {
      "get" : {
        "tags" : [ "PopularTimes" ],
        "summary" : "populartimes",
        "description" : "populartimes",
        "operationId" : "populartimes",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PopularTimesResponse"
            }
          }
        }
      }
    },
    "/populartimes/spid/{spid}/cc/{cc}" : {
      "get" : {
        "tags" : [ "PopularTimes" ],
        "summary" : "populartimes",
        "description" : "populartimes",
        "operationId" : "populartimesCountry",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "spid",
          "in" : "path",
          "description" : "ServicePointID",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "cc",
          "in" : "path",
          "description" : "CountryCode",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/PopularTimesResponse"
            }
          }
        }
      }
    },
...
    "/volume/{from}/{to}/service/{service}/consignor/{consignor}" : {
      "get" : {
        "tags" : [ "Volume" ],
        "summary" : "volume",
        "description" : "volume",
        "operationId" : "volume",
        "produces" : [ "text/plain" ],
        "parameters" : [ {
          "name" : "from",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "to",
          "in" : "path",
          "required" : true,
          "type" : "integer",
          "format" : "int32"
        }, {
          "name" : "service",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "consignor",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    }
  },
  "definitions" : {
    "ApiError" : {
      "type" : "object",
      "required" : [ "faults" ],
      "properties" : {
        "faults" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Fault"
          }
        }
      },
      "description" : "Model for API errors"
    },
...
    "Fault" : {
      "type" : "object",
      "required" : [ "explanationText", "faultCode" ],
      "properties" : {
        "faultCode" : {
          "type" : "string"
        },
        "explanationText" : {
          "type" : "string"
        },
        "paramValues" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/ParamValues"
          }
        }
      },
      "description" : "Model for API Fault"
    },
...
    "ParamValues" : {
      "type" : "object",
      "required" : [ "param", "value" ],
      "properties" : {
        "param" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        }
      },
      "description" : "Key/Value information about errors"
    },
    "PopularTimesResponse" : {
      "type" : "object",
      "properties" : {
        "error" : {
          "type" : "string"
        },
        "apiError" : {
          "$ref" : "#/definitions/ApiError"
        },
        "name" : {
          "type" : "string"
        },
        "date" : {
          "type" : "string"
        },
        "dateFrom" : {
          "type" : "string",
          "format" : "date-time"
        },
        "dateTo" : {
          "type" : "string",
          "format" : "date-time"
        },
        "servicePointId" : {
          "type" : "string"
        },
        "popularTimes" : {
          "type" : "array",
          "items" : {
            "type" : "object",
            "additionalProperties" : {
              "type" : "object"
            }
          }
        },
        "weekAverage" : {
          "type" : "integer",
          "format" : "int64"
        },
        "countryCode" : {
          "type" : "string"
        }
      }
    }
  }
}